Extract reusable code from OutputModel
