+++ nbdime-web/src/diff/model/index.ts
@@ -6,4 +6,5 @@ export * from './common';
 export * from './immutable';
 export * from './notebook';
 export * from './output';
+export * from './renderable';
 export * from './string';
